// tony/Week10/src/types/movie.ts
export interface Movie {
	id: number;
	title: string;
	poster_path: string | null;
	backdrop_path: string | null;
	overview: string;
	release_date: string;
	vote_average: number;
	vote_count: number;
	genre_ids: number[];
	adult: boolean;
	original_language: string;
	popularity: number;
}

export interface MovieSearchResponse {
	page: number;
	results: Movie[];
	total_pages: number;
	total_results: number;
}

export type Language = "ko-KR" | "en-US" | "ja-JP";

export interface SearchParams {
	query: string;
	includeAdult: boolean;
	language: Language;
}